move view and form to /member/ app
