+++ saskatoon/member/views.py
@@ -1,12 +1,16 @@
 from django.contrib.messages.views import SuccessMessageMixin
 from django.contrib.auth.mixins import PermissionRequiredMixin
+from django.contrib.auth import views as auth_views
 from django.views.generic import CreateView, UpdateView
 from django.urls import reverse_lazy
 from django.utils.translation import gettext_lazy as _
+from django.contrib import messages
+from django.urls import reverse_lazy
 from .models import Person, Organization
 from harvest.models import Property
 from .forms import ( PersonCreateForm, PersonUpdateForm,
-                     OrganizationCreateForm, OrganizationForm )
+                     OrganizationCreateForm, OrganizationForm,
+                     PasswordChangeForm)
 
 
 class PersonCreateView(PermissionRequiredMixin, SuccessMessageMixin, CreateView):
@@ -113,3 +117,16 @@ class OrganizationUpdateView(PermissionRequiredMixin, SuccessMessageMixin, Updat
         context['title'] = _("Organization Update")
         context['cancel_url'] = reverse_lazy('beneficiary-list')
         return context
+
+
+class PasswordChangeView(auth_views.PasswordChangeView):
+    template_name = 'registration/change_password.html'
+    form_class = PasswordChangeForm
+
+    def get_context_data(self, **kwargs):
+        context = super(PasswordChangeView, self).get_context_data(**kwargs)
+        return context
+
+    def get_success_url(self):
+        messages.success(self.request, _("Password successfully changed!"))
+        return reverse_lazy('home')
