+++ saskatoon/sitebase/urls.py
@@ -30,10 +30,6 @@ urlpatterns = [
          views.VolunteerWaiverPDFView.as_view(),
          name='volunteer-waiver'),
 
-    path('change_password/',
-         views.PasswordChangeView.as_view(),
-         name ='change_password'),
-
     path('reset_password/',
          auth_views.PasswordResetView.as_view(),
          name ='reset_password'),
